Updated Atomic class to deal with float and double types.

+++ extras/Jucer (experimental)/Source/Project/jucer_Project.cpp
@@ -103,7 +103,7 @@ void Project::setMissingDefaultValues()
         setTitle ("Juce Project");
 
     if (! projectRoot.hasProperty (Ids::projectType))
-        getProjectType() = (int) application;
+        getProjectType() = application;
 
     if (! projectRoot.hasProperty (Ids::version))
         getVersion() = "1.0.0";
